refactor file_data to be clearer
